+++ src/components/HistoryTable/hooks/useTestResults.test.tsx
@@ -1,30 +1,25 @@
 import { act, renderHook } from "@testing-library/react-hooks";
 import { GET_TASK_TEST_SAMPLE } from "gql/queries";
 import { TestStatus } from "types/history";
+import { useHistoryTable } from "../HistoryTableContext";
 import { mainlineCommitData } from "../testData";
 import { rowType } from "../types";
-import { useHistoryTableTestHook, ProviderWrapper } from "./test-utils";
-import useTestResultsActual from "./useTestResults";
+import { ProviderWrapper } from "./test-utils";
+import useTestResults from "./useTestResults";
 
 describe("useTestResults", () => {
   it("should return an empty map when nothing is loaded", () => {
-    const { result } = renderHook(
-      () => useHistoryTableTestHook(useTestResultsActual, [0]),
-      {
-        wrapper: ({ children }) => ProviderWrapper({ children }),
-      }
-    );
+    const { result } = renderHook(() => useMergedTestHook(0), {
+      wrapper: ({ children }) => ProviderWrapper({ children }),
+    });
     expect(result.current.hookResponse).toStrictEqual({
       getTaskMetadata: expect.any(Function),
     });
   });
   it("should return the default state when there is no valid data for a row", () => {
-    const { result } = renderHook(
-      () => useHistoryTableTestHook(useTestResultsActual, [0]),
-      {
-        wrapper: ({ children }) => ProviderWrapper({ children }),
-      }
-    );
+    const { result } = renderHook(() => useMergedTestHook(0), {
+      wrapper: ({ children }) => ProviderWrapper({ children }),
+    });
     expect(
       result.current.hookResponse.getTaskMetadata(
         "evergreen_ubuntu1604_dist_d4cf298cf0b2536fb3bff875775b93a9ceafb75c_21_09_02_14_20_04"
@@ -36,12 +31,9 @@ describe("useTestResults", () => {
     });
   });
   it("should not attempt to fetch data for non commit rows", () => {
-    const { result } = renderHook(
-      () => useHistoryTableTestHook(useTestResultsActual, [0]),
-      {
-        wrapper: ({ children }) => ProviderWrapper({ children, mocks }),
-      }
-    );
+    const { result } = renderHook(() => useMergedTestHook(0), {
+      wrapper: ({ children }) => ProviderWrapper({ children, mocks }),
+    });
     expect(
       result.current.hookResponse.getTaskMetadata("some_id")
     ).toMatchObject({
@@ -66,7 +58,7 @@ describe("useTestResults", () => {
   });
   it("should return all matching test results when there are no filters applied and the row is a commit", async () => {
     const { result, waitForNextUpdate } = renderHook(
-      () => useHistoryTableTestHook(useTestResultsActual, [1]),
+      () => useMergedTestHook(1),
       {
         wrapper: ({ children }) => ProviderWrapper({ children, mocks }),
       }
@@ -100,7 +92,7 @@ describe("useTestResults", () => {
   });
   it("should return all matching test results when there are matching filters applied and the row is a commit", async () => {
     const { result, waitForNextUpdate } = renderHook(
-      () => useHistoryTableTestHook(useTestResultsActual, [1]),
+      () => useMergedTestHook(1),
       {
         wrapper: ({ children }) => ProviderWrapper({ children, mocks }),
       }
@@ -141,7 +133,7 @@ describe("useTestResults", () => {
   });
   it("should not return matching test results when there are non matching filters applied and the row is a commit", async () => {
     const { result, waitForNextUpdate } = renderHook(
-      () => useHistoryTableTestHook(useTestResultsActual, [1]),
+      () => useMergedTestHook(1),
       {
         wrapper: ({ children }) => ProviderWrapper({ children, mocks }),
       }
@@ -176,6 +168,58 @@ describe("useTestResults", () => {
       inactive: true,
       label: "0 / 1 Failing Tests",
       failingTests: [],
+    });
+  });
+});
+
+type UseMergedTestHookType = (
+  args: Parameters<typeof useTestResults>[0]
+) => {
+  hookResponse: ReturnType<typeof useTestResults>;
+  historyTable: ReturnType<typeof useHistoryTable>;
+};
+/** useMergedTestHook combines the useTestResults and useHistoryTable hooks together
+ * and combines them into a shared hook which can be rendered under the same wrapper context
+ * and can be used together */
+const useMergedTestHook: UseMergedTestHookType = (args) => {
+  const hookResponse = useTestResults(args);
+  const historyTable = useHistoryTable();
+
+  return {
+    hookResponse,
+    historyTable,
+  };
+};
+
+// This is a sanity check to ensure the useMergedHookRender hook is working as expected
+describe("useMergedHookRender - sanity check", () => {
+  it("should return the correct hooks", () => {
+    const { result } = renderHook(() => useMergedTestHook(0), {
+      wrapper: ProviderWrapper,
+    });
+    expect(result.current.hookResponse).toStrictEqual({
+      getTaskMetadata: expect.any(Function),
+    });
+    expect(result.current.historyTable).toStrictEqual({
+      processedCommitCount: 0,
+      fetchNewCommit: expect.any(Function),
+      getItem: expect.any(Function),
+      isItemLoaded: expect.any(Function),
+      getItemHeight: expect.any(Function),
+      toggleRowSizeAtIndex: expect.any(Function),
+      hasNextPage: false,
+      hasPreviousPage: false,
+      historyTableFilters: [],
+      setHistoryTableFilters: expect.any(Function),
+      processedCommits: [],
+      visibleColumns: [],
+      addColumns: expect.any(Function),
+      nextPage: expect.any(Function),
+      previousPage: expect.any(Function),
+      currentPage: 0,
+      pageCount: 0,
+      columnLimit: 7,
+      commitCount: 10,
     });
   });
 });
